refactor: implement ADR-061 aggregation phase
