+++ changelog.d/20260825_adr061_phase1.md
@@ -1,0 +1,6 @@
+### Changed
+
+- Implemented ADR-061 Phase 1 by moving multi-target aggregation into the
+  `abicheck.workflows.aggregate` responsibility package, retaining the
+  documented `abicheck.aggregate` imports as a thin compatibility facade and
+  routing the CLI through the canonical workflow and report projection owners.
