+++ docs/contribute/adr/index.md
@@ -155,4 +155,4 @@ against the code as unverified, regardless of how confident it reads.
 | [058](058-native-compatibility-agent-skills.md) | Native Compatibility Agent Skills — User-Task-First Domain Layer | Accepted — partially implemented (G36 P0.1–P0.3/P0.6–P0.9: the generator and gates shipped; P0.4/P0.5 product-surface items and all of P1 remain). Portfolio reset to one skill (2026-08-20 amendment, superseding the 2026-08-11 four-skill freeze): `review-native-library-change` (renamed from `native-binary-compatibility-review`) is the sole published skill, an unvalidated internal candidate; the other three are removed from `skills-src/` and every generated tree, recoverable from git history. A second, same-date amendment ("PR 2") then rewrote that skill's workflow content — customer-outcome framing, a ten-step decision procedure, an integrated named-consumer branch, a narrowed v0.1 validated scope, a structured decision-report contract — still unvalidated. A third, same-date amendment ("PR 3.5") renamed it again to `check-abi-compatibility` (a user-outcome name rather than a mechanism name) and recorded, as design intent only, that PR 4's external-distribution deliverable should be an npm/npx-installable package published from this repository rather than a separate distribution repo. A fourth, same-date amendment ("PR 3") landed the full G37 evaluation corpus (12 scenarios) and a real 48-run pilot under the new name; its dominant finding is a harness confound (a 12-turn budget cutting off 31% of runs, asymmetrically by arm) rather than a skill-quality result — see `agent-evals/skills/pilot-results/README.md`. A fifth amendment ("Harbor task battery") added a generated [Harbor](https://www.harborframework.com) task directory per scenario (`agent-evals/skills/harbor/tasks/`) — schema-validated against the real `harbor` package and end-to-end verified for Category A reference solutions, but never run through an actual Harbor trial (no working container/sandbox runtime in this environment). A sixth, same-day amendment ("Harbor made canonical") then decided Harbor is the surface for all new scenario/trial work going forward; `runners/claude_code.py` is kept only as the historical record of the existing pilot — see `agent-evals/skills/harbor/CLAUDE.md`. Still unvalidated; see [G36](../plans/g36-native-compatibility-agent-skills.md) / [G37](../plans/g37-agent-skill-quality-evaluation.md) |
 | [059](059-compressed-snapshot-storage.md) | Compressed Snapshot Storage Envelope | Accepted — implemented (core snapshot I/O, `dump`, `compare`/`scan --against`/Python API, the snapshot cache, `actions/baseline`, the root Action's `dump`-mode `snapshot-compression` input, compressed-release-asset auto-fetch, and both publish workflows' `snapshot-compression` input; `resolve-baseline` needed no changes at all — already transparent); baseline-set manifest v2, a deterministic `.tar.zst` packager, and the wider docs sweep deferred, see the ADR's own "What this ADR does not (yet) close" |
 | [060](060-synthetic-consumer-compile-probe-deferral.md) | Synthetic-Consumer Compile-Probe Layer — Deferred | Accepted — not implemented (decision to defer), and no follow-up phase is scheduled; see [G31](../plans/g31-header-graph-default-on-followup.md) Phase D |
-| [061](061-responsibility-package-architecture.md) | Responsibility-Package Architecture and Flat-Namespace Migration | Accepted — Phase 0 implemented (ownership guidance, stable module graph, measured no-growth debt ledger, focused architecture gate, and PR-profile wiring); Phases 1-5 remain incremental |
+| [061](061-responsibility-package-architecture.md) | Responsibility-Package Architecture and Flat-Namespace Migration | Accepted — Phases 0-1 implemented (architecture enforcement plus the aggregation workflow migration and compatibility facade); Phases 2-5 remain incremental |
